requirements.txt
openai==1.25.0
fastapi==0.111.0
uvicorn==0.29.0
bs4==0.0.2
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
weaviate-client==4.5.7

